==> 04_2_sensitivity_token512.ipynb <==
from transformers import BertModel, BertTokenizer
import torch
import torch.nn as nn
import os
import pandas as pd
import numpy as np
import tiktoken

df = pd.read_csv("002_data/data_wide.csv", index_col=0)
df.head(3)


time_columns = [col for col in df.columns if col.startswith('time')]

df[time_columns] = df[time_columns].astype(str)

def combine(row):
    combined_values = ' '.join(row[col] for col in time_columns)
    return combined_values

df['combined'] = df.apply(combine, axis=1)

embedding_encoding = "cl100k_base"
encoding = tiktoken.get_encoding(embedding_encoding)
df["n_tokens"] = df.combined.apply(lambda x: len(encoding.encode(x)))

df = df.drop(columns=[col for col in df.columns if col.startswith('time')])
df.to_csv("002_data/data_wide_embedding_initial.csv")

df = pd.read_csv("002_data/data_wide_embedding_initial.csv", index_col=0)
df.head(2)

################### GPT ##############################

import tiktoken
from openai import OpenAI
import ast

# model: text-embedding-3-small
# default embedding dimension: 1536

from openai import OpenAI
api_key = ''
client = OpenAI(api_key=api_key)

def get_embedding(text, model="text-embedding-3-small", max_tokens=512):  # set the maximum equal to MOMENT
    text = text.replace("\n", " ")
    encoding = tiktoken.get_encoding("cl100k_base")
    tokens = encoding.encode(text)[:max_tokens]
    text = encoding.decode(tokens)
    return client.embeddings.create(input=[text], model=model).data[0].embedding

embedding_model = "text-embedding-3-small"
embedding_encoding = "cl100k_base"
max_tokens = 512   

df_gpt = df.copy() 
df_gpt["embedding"] = df_gpt.combined.apply(lambda x: get_embedding(x, model=embedding_model, max_tokens=max_tokens))
df_gpt["embedding"] = df_gpt["embedding"].apply(lambda x: str(x))

df_gpt = df_gpt[[col for col in df_gpt.columns if not col.startswith('n_tokens') and not col.startswith('time')]]
df_gpt.to_csv("009_untitled/trunc_wide_embedding_gpt1536.csv")
df_gpt.head(2)
